updated dashboard page components ui and ensure responsiveness
