+++ src/app/(dashboard)/dashboard/admin/manage-order/page.jsx
@@ -217,7 +217,7 @@ export default function ManageOrders() {
 
   return (
     <div className="p-4 sm:p-6 bg-gray-50 dark:bg-gray-900 min-h-screen transition-colors duration-300">
-      <div className="max-w-7xl mx-auto">
+      <div className="w-full">
         {/* Header */}
         <div className="mb-8 text-center sm:text-left">
           <h1 className="text-3xl font-extrabold text-gray-900 dark:text-white">
